refactor: rename `tailwind-config` for `tailwind`
