+++ apps/finance/tailwind.config.ts
@@ -1,6 +1,6 @@
 import type { Config } from 'tailwindcss';
 
-import { aria, scrollbars, theme } from '@stack/tailwind-config';
+import { aria, scrollbars, theme } from '@stack/tailwind';
 
 export default {
   content: ['./src/**/*.{html,js,svelte,ts}'],
