+++ packages/actions/package.json
@@ -24,7 +24,7 @@
     "@stack/utils": "workspace:*"
   },
   "devDependencies": {
-    "@stack/tailwind-config": "workspace:*",
+    "@stack/tailwind": "workspace:*",
     "@stack/typescript-config": "workspace:*",
     "@sveltejs/kit": "2.16.1",
     "@sveltejs/package": "2.3.9",
