+++ packages/layouts/package.json
@@ -31,7 +31,7 @@
     "tailwind-merge": "3.0.1"
   },
   "devDependencies": {
-    "@stack/tailwind-config": "workspace:*",
+    "@stack/tailwind": "workspace:*",
     "@sveltejs/kit": "2.16.1",
     "@sveltejs/vite-plugin-svelte": "5.0.3",
     "svelte": "5.19.6"
