+++ packages/layouts/tailwind.config.ts
@@ -1,6 +1,6 @@
 import type { Config } from 'tailwindcss';
 
-import { aria, scrollbars } from '@stack/tailwind-config';
+import { aria, scrollbars } from '@stack/tailwind';
 
 export default {
   plugins: [scrollbars(), aria()]
