+++ packages/tailwind/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@stack/tailwind-config",
+  "name": "@stack/tailwind",
   "private": true,
   "version": "0.0.0",
   "exports": {
