+++ packages/tailwind/tsconfig.json
