+++ packages/ui/package.json
@@ -29,7 +29,7 @@
     "tailwind-merge": "3.0.1"
   },
   "devDependencies": {
-    "@stack/tailwind-config": "workspace:*",
+    "@stack/tailwind": "workspace:*",
     "@stack/typescript-config": "workspace:*",
     "@sveltejs/kit": "2.16.1",
     "@sveltejs/package": "2.3.9",
