+++ turbo/generators/templates/app/package.json.hbs
@@ -9,7 +9,7 @@ dev", "dev:app": "vite dev", "lint": "eslint . --max-warnings=0", "lint:fix":
 "playwright test", "test:unit": "vitest", "type-check": "tsc --noEmit" },
 "dependencies": { "@supabase/ssr": "0.5.1", "@supabase/supabase-js": "2.45.4",
 "@tanstack/svelte-query": "5.59.0", "lucide-svelte": "0.447.0" },
-"devDependencies": { "@playwright/test": "1.47.2", "@stack/tailwind-config":
+"devDependencies": { "@playwright/test": "1.47.2", "@stack/tailwind":
 "workspace:*", "@stack/typescript-config": "workspace:*",
 "@sveltejs/adapter-auto": "3.2.5", "@sveltejs/adapter-node": "5.2.5",
 "@sveltejs/kit": "2.6.2", "@sveltejs/vite-plugin-svelte": "4.0.0-next.6",
